refactor: 오류 응답 DTO를 core 계층으로 이동
